+++ src/lib.rs
@@ -57,16 +57,15 @@ pub mod jws;
 
 use errors::{Error, ValidationError};
 
-/// A part of the JWT: header and claims specifically
-/// Allows converting from/to struct with base64
-trait Part {
+/// A "part" of the compact representation of JWT/JWS/JWE. Parts are base64 encoded and separated by periods
+trait CompactPart {
     type Encoded: AsRef<str>;
 
     fn from_base64<B: AsRef<[u8]>>(encoded: B) -> Result<Self, Error> where Self: Sized;
     fn to_base64(&self) -> Result<Self::Encoded, Error>;
 }
 
-impl<T> Part for T
+impl<T> CompactPart for T
     where T: Serialize + Deserialize
 {
     type Encoded = String;
@@ -530,228 +529,6 @@ impl<T> Clone for ClaimsSet<T>
     }
 }
 
-/// A JWT that can be encoded/decoded
-///
-/// The serialization/deserialization is handled by serde. Before you transport the token, make sure you
-/// turn it into the encoded form first.
-///
-/// # Examples
-/// ## Encoding and decoding with HS256
-///
-/// ```
-/// extern crate biscuit;
-/// extern crate serde;
-/// #[macro_use]
-/// extern crate serde_derive;
-/// extern crate serde_json;
-///
-/// use std::str::FromStr;
-/// use biscuit::*;
-/// use biscuit::jws::*;
-///
-/// # fn main() {
-///
-/// // Define our own private claims
-/// #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-/// struct PrivateClaims {
-///     company: String,
-///     department: String,
-/// }
-///
-/// let expected_token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.\
-///                         eyJpc3MiOiJodHRwczovL3d3dy5hY21lLmNv\
-///                         bS8iLCJzdWIiOiJKb2huIERvZSIsImF1ZCI6I\
-///                         mh0dHM6Ly9hY21lLWN1c3RvbWVyLmNvbS8iLC\
-///                         JuYmYiOjEyMzQsImNvbXBhbnkiOiJBQ01FIiwi\
-///                         ZGVwYXJ0bWVudCI6IlRvaWxldCBDbG\
-///                         VhbmluZyJ9.dnx1OmRZSFxjCD1ivy4lveTT-sxay5Fq6vY6jnJvqeI";
-///
-/// let expected_claims = ClaimsSet::<PrivateClaims> {
-///     registered: RegisteredClaims {
-///         issuer: Some(FromStr::from_str("https://www.acme.com").unwrap()),
-///         subject: Some(FromStr::from_str("John Doe").unwrap()),
-///         audience:
-///             Some(SingleOrMultiple::Single(FromStr::from_str("htts://acme-customer.com").unwrap())),
-///         not_before: Some(1234.into()),
-///         ..Default::default()
-///     },
-///     private: PrivateClaims {
-///         department: "Toilet Cleaning".to_string(),
-///         company: "ACME".to_string(),
-///     },
-/// };
-///
-/// let expected_jwt = JWT::new_decoded(Header {
-///                                         algorithm: Algorithm::HS256,
-///                                         ..Default::default()
-///                                     },
-///                                     expected_claims.clone());
-///
-/// let token = expected_jwt
-///     .into_encoded(Secret::Bytes("secret".to_string().into_bytes())).unwrap();
-/// let token = serde_json::to_string(&token).unwrap();
-/// assert_eq!(format!("\"{}\"", expected_token), token);
-/// // Now, send `token` to your clients
-///
-/// // ... some time later, we get token back!
-///
-/// let token = serde_json::from_str::<JWT<PrivateClaims>>(&token).unwrap();
-/// let token = token.into_decoded(Secret::Bytes("secret".to_string().into_bytes()),
-///     Algorithm::HS256).unwrap();
-/// assert_eq!(*token.claims_set().unwrap(), expected_claims);
-/// # }
-/// ```
-#[derive(Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum JWT<T: Serialize + Deserialize> {
-    /// Decoded form of the JWT. **DO NOT SERIALIZE THIS AND SEND TO YOUR CLIENTS.**.
-    /// This variant cannot be serialized or deserialized and will return an error.
-    #[serde(skip_serializing)]
-    #[serde(skip_deserializing)]
-    Decoded {
-        /// Embedded header
-        header: jws::Header,
-        /// Claims sets, including registered and private claims
-        claims_set: ClaimsSet<T>,
-    },
-    /// Encoded and (optionally) signed JWT. Use this form to send to your clients
-    Encoded(String),
-}
-
-/// Used in decode: takes the result of a rsplit and ensure we only get 2 parts
-/// Errors if we don't
-macro_rules! expect_two {
-    ($iter:expr) => {{
-        let mut i = $iter; // evaluate the expr
-        match (i.next(), i.next(), i.next()) {
-            (Some(first), Some(second), None) => Ok((first, second)),
-            _ => Err(Error::ValidationError(ValidationError::InvalidToken))
-        }
-    }}
-}
-
-impl<T: Serialize + Deserialize> JWT<T> {
-    /// New decoded JWT
-    pub fn new_decoded(header: jws::Header, claims_set: ClaimsSet<T>) -> Self {
-        JWT::Decoded {
-            header: header,
-            claims_set: claims_set,
-        }
-    }
-
-    /// New encoded JWT
-    pub fn new_encoded(token: &str) -> Self {
-        JWT::Encoded(token.to_string())
-    }
-
-    /// Consumes self and convert into encoded form. If the token is already encoded,
-    /// this is a no-op.
-    // TODO: Is the no-op dangerous? What if the secret between the previous encode and this time is different?
-    pub fn into_encoded(self, secret: jws::Secret) -> Result<Self, Error> {
-        match self {
-            JWT::Encoded(_) => Ok(self),
-            JWT::Decoded { .. } => self.encode(secret),
-        }
-    }
-
-    /// Encode the JWT passed and sign the payload using the algorithm from the header and the secret
-    /// The secret is dependent on the signing algorithm
-    pub fn encode(&self, secret: jws::Secret) -> Result<Self, Error> {
-        match *self {
-            JWT::Decoded { ref header, ref claims_set } => {
-                let encoded_header = header.to_base64()?;
-                let encoded_claims = claims_set.to_base64()?;
-                let payload = [&*encoded_header, &*encoded_claims].join(".");
-                let signature = header.algorithm
-                    .sign(payload.as_bytes(), secret)?
-                    .as_slice()
-                    .to_base64(base64::URL_SAFE);
-
-                Ok(JWT::Encoded([payload, signature].join(".")))
-            }
-            JWT::Encoded(_) => Err(Error::UnsupportedOperation),
-        }
-    }
-
-
-    /// Consumes self and convert into decoded form, verifying the signature, if any.
-    /// If the token is already decoded, this is a no-op.AsMut
-    // TODO: Is the no-op dangerous? What if the secret between the previous decode and this time is different?
-    pub fn into_decoded(self, secret: jws::Secret, algorithm: jws::Algorithm) -> Result<Self, Error> {
-        match self {
-            JWT::Encoded(_) => self.decode(secret, algorithm),
-            JWT::Decoded { .. } => Ok(self),
-        }
-    }
-
-    /// Decode a token into the JWT struct and verify its signature
-    /// If the token or its signature is invalid, it will return an error
-    pub fn decode(&self, secret: jws::Secret, algorithm: jws::Algorithm) -> Result<Self, Error> {
-        match *self {
-            JWT::Decoded { .. } => Err(Error::UnsupportedOperation),
-            JWT::Encoded(ref token) => {
-                // Check that there are only two parts
-                let (signature, payload) = expect_two!(token.rsplitn(2, '.'))?;
-                let signature: Vec<u8> = signature.from_base64()?;
-
-                if !algorithm.verify(signature.as_ref(), payload.as_ref(), secret)? {
-                    Err(ValidationError::InvalidSignature)?;
-                }
-
-                let (claims, header) = expect_two!(payload.rsplitn(2, '.'))?;
-
-                let header = jws::Header::from_base64(header)?;
-                if header.algorithm != algorithm {
-                    Err(ValidationError::WrongAlgorithmHeader)?;
-                }
-                let decoded_claims = ClaimsSet::<T>::from_base64(claims)?;
-
-                Ok(Self::new_decoded(header, decoded_claims))
-            }
-        }
-    }
-
-    /// Convenience method to extract the encoded string from an encoded JWT
-    pub fn encoded(&self) -> Result<&str, Error> {
-        match *self {
-            JWT::Decoded { .. } => Err(Error::UnsupportedOperation),
-            JWT::Encoded(ref encoded) => Ok(encoded),
-        }
-    }
-
-    /// Convenience method to extract the claims set from a decoded JWT
-    pub fn claims_set(&self) -> Result<&ClaimsSet<T>, Error> {
-        match *self {
-            JWT::Decoded { ref claims_set, .. } => Ok(claims_set),
-            JWT::Encoded(_) => Err(Error::UnsupportedOperation),
-        }
-    }
-
-    /// Convenience method to extract the header from a decoded JWT
-    pub fn header(&self) -> Result<&jws::Header, Error> {
-        match *self {
-            JWT::Decoded { ref header, .. } => Ok(header),
-            JWT::Encoded(_) => Err(Error::UnsupportedOperation),
-        }
-    }
-}
-
-impl<T> Clone for JWT<T>
-    where T: Serialize + Deserialize + Clone
-{
-    fn clone(&self) -> Self {
-        match *self {
-            JWT::Decoded { ref header, ref claims_set } => {
-                JWT::Decoded {
-                    header: (*header).clone(),
-                    claims_set: (*claims_set).clone(),
-                }
-            }
-            JWT::Encoded(ref encoded) => JWT::Encoded((*encoded).clone()),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     extern crate serde_test;
@@ -765,7 +542,6 @@ mod tests {
     use self::serde_test::{Token, assert_tokens, assert_ser_tokens_error};
 
     use super::*;
-    use jws::{Algorithm, Header, Secret};
 
     #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
     struct PrivateClaims {
@@ -1086,203 +862,6 @@ mod tests {
         };
 
         serde_json::to_string(&claim).unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "the enum variant JWT::Decoded cannot be serialized")]
-    fn decoded_jwt_cannot_be_serialized() {
-        let expected_claims = ClaimsSet::<PrivateClaims> {
-            registered: RegisteredClaims {
-                issuer: Some(not_err!(FromStr::from_str("https://www.acme.com/"))),
-                subject: Some(not_err!(FromStr::from_str("John Doe"))),
-                audience: Some(SingleOrMultiple::Single(not_err!(FromStr::from_str("htts://acme-customer.com/")))),
-                not_before: Some(1234.into()),
-                ..Default::default()
-            },
-            private: PrivateClaims {
-                department: "Toilet Cleaning".to_string(),
-                company: "ACME".to_string(),
-            },
-        };
-
-        let biscuit = JWT::new_decoded(Header { algorithm: Algorithm::None, ..Default::default() },
-                                       expected_claims.clone());
-        serde_json::to_string(&biscuit).unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "data did not match any variant of untagged enum JWT")]
-    fn decoded_jwt_cannot_be_deserialized() {
-        let json = r#"{"header":{"alg":"none","typ":"JWT"},
-                       "claims_set":{"iss":"https://www.acme.com/","sub":"John Doe",
-                                     "aud":"htts://acme-customer.com","nbf":1234,
-                                     "company":"ACME","department":"Toilet Cleaning"}}"#;
-        serde_json::from_str::<JWT<PrivateClaims>>(json).unwrap();
-    }
-
-    #[test]
-    fn round_trip_none() {
-        let expected_token = "eyJhbGciOiJub25lIiwidHlwIjoiSldUIn0.\
-                              eyJpc3MiOiJodHRwczovL3d3dy5hY21lLmNvbS8iLCJzdWIiOiJKb2huIERvZSIsImF1ZCI6Imh0dHM6Ly9\
-                              hY21lLWN1c3RvbWVyLmNvbS8iLCJuYmYiOjEyMzQsImNvbXBhbnkiOiJBQ01FIiwiZGVwYXJ0bWVudCI6Il\
-                              RvaWxldCBDbGVhbmluZyJ9.";
-
-        let expected_claims = ClaimsSet::<PrivateClaims> {
-            registered: RegisteredClaims {
-                issuer: Some(not_err!(FromStr::from_str("https://www.acme.com"))),
-                subject: Some(not_err!(FromStr::from_str("John Doe"))),
-                audience: Some(SingleOrMultiple::Single(not_err!(FromStr::from_str("htts://acme-customer.com")))),
-                not_before: Some(1234.into()),
-                ..Default::default()
-            },
-            private: PrivateClaims {
-                department: "Toilet Cleaning".to_string(),
-                company: "ACME".to_string(),
-            },
-        };
-
-        let expected_jwt = JWT::new_decoded(Header { algorithm: Algorithm::None, ..Default::default() },
-                                            expected_claims.clone());
-        let token = not_err!(expected_jwt.into_encoded(Secret::None));
-        assert_eq!(expected_token, not_err!(token.encoded()));
-
-        let biscuit = not_err!(token.into_decoded(Secret::None, Algorithm::None));
-        let actual_claims = not_err!(biscuit.claims_set());
-        assert_eq!(expected_claims, *actual_claims);
-    }
-
-    #[test]
-    fn round_trip_hs256() {
-        let expected_token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.\
-                              eyJpc3MiOiJodHRwczovL3d3dy5hY21lLmNvbS8iLCJzdWIiOiJKb2huIERvZSIsImF1ZCI6Imh0dHM6Ly9hY21lL\
-                              WN1c3RvbWVyLmNvbS8iLCJuYmYiOjEyMzQsImNvbXBhbnkiOiJBQ01FIiwiZGVwYXJ0bWVudCI6IlRvaWxldCBDbG\
-                              VhbmluZyJ9.dnx1OmRZSFxjCD1ivy4lveTT-sxay5Fq6vY6jnJvqeI";
-
-        let expected_claims = ClaimsSet::<PrivateClaims> {
-            registered: RegisteredClaims {
-                issuer: Some(not_err!(FromStr::from_str("https://www.acme.com"))),
-                subject: Some(not_err!(FromStr::from_str("John Doe"))),
-                audience: Some(SingleOrMultiple::Single(not_err!(FromStr::from_str("htts://acme-customer.com")))),
-                not_before: Some(1234.into()),
-                ..Default::default()
-            },
-            private: PrivateClaims {
-                department: "Toilet Cleaning".to_string(),
-                company: "ACME".to_string(),
-            },
-        };
-
-        let expected_jwt = JWT::new_decoded(Header { algorithm: Algorithm::HS256, ..Default::default() },
-                                            expected_claims.clone());
-        let token = not_err!(expected_jwt.into_encoded(Secret::Bytes("secret".to_string().into_bytes())));
-        assert_eq!(expected_token, not_err!(token.encoded()));
-
-        let biscuit = not_err!(token.into_decoded(Secret::Bytes("secret".to_string().into_bytes()),
-                                                  Algorithm::HS256));
-        assert_eq!(expected_claims, *not_err!(biscuit.claims_set()));
-    }
-
-    #[test]
-    fn round_trip_rs256() {
-        let expected_token = "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.\
-                              eyJpc3MiOiJodHRwczovL3d3dy5hY21lLmNvbS8iLCJzdWIiOiJKb2huIERvZSIsImF1ZCI6Imh0dHM6Ly9hY21lL\
-                              WN1c3RvbWVyLmNvbS8iLCJuYmYiOjEyMzQsImNvbXBhbnkiOiJBQ01FIiwiZGVwYXJ0bWVudCI6IlRvaWxldCBDbG\
-                              VhbmluZyJ9.THHNGg4AIq2RT30zecAD41is6j1ffGRn6GdK6cpl08esHufG5neJOMTO1fONVykOFgCaJw9jLP7GCd\
-                              YumsMKU3434QAQyvLCPklHQWE7VcSFSdsf7skcvuvwPtkMWCGrzFK7seVv9OiJzjNzoeyS2d8io7wviFqkpcXwOVZ\
-                              W4ArP5katX4nIoXlwWfcK82E6MacSIL2uq_ha6yL2z7trq3dSszSnUevlWKq-9FIFk11XwToMTmGubkWyGk-k-dfH\
-                              AXwnS1hADXkwSAemWoCG98v6zFtTZHOOAPnB09acEKVtVRFKZQa3V2IpdsHtRoPJU5pFgCXi8VRebHJm99yTXw";
-
-        let expected_claims = ClaimsSet::<PrivateClaims> {
-            registered: RegisteredClaims {
-                issuer: Some(not_err!(FromStr::from_str("https://www.acme.com"))),
-                subject: Some(not_err!(FromStr::from_str("John Doe"))),
-                audience: Some(SingleOrMultiple::Single(not_err!(FromStr::from_str("htts://acme-customer.com")))),
-                not_before: Some(1234.into()),
-                ..Default::default()
-            },
-            private: PrivateClaims {
-                department: "Toilet Cleaning".to_string(),
-                company: "ACME".to_string(),
-            },
-        };
-        let private_key = Secret::rsa_keypair_from_file("test/fixtures/rsa_private_key.der").unwrap();
-
-        let expected_jwt = JWT::new_decoded(Header { algorithm: Algorithm::RS256, ..Default::default() },
-                                            expected_claims.clone());
-        let token = not_err!(expected_jwt.into_encoded(private_key));
-        assert_eq!(expected_token, not_err!(token.encoded()));
-
-        let public_key = Secret::public_key_from_file("test/fixtures/rsa_public_key.der").unwrap();
-        let biscuit = not_err!(token.into_decoded(public_key, Algorithm::RS256));
-        assert_eq!(expected_claims, *not_err!(biscuit.claims_set()));
-    }
-
-    #[test]
-    fn encode_with_additional_header_fields() {
-        let expected_claims = ClaimsSet::<PrivateClaims> {
-            registered: RegisteredClaims {
-                issuer: Some(not_err!(FromStr::from_str("https://www.acme.com"))),
-                subject: Some(not_err!(FromStr::from_str("John Doe"))),
-                audience: Some(SingleOrMultiple::Single(not_err!(FromStr::from_str("htts://acme-customer.com")))),
-                not_before: Some(1234.into()),
-                ..Default::default()
-            },
-            private: PrivateClaims {
-                department: "Toilet Cleaning".to_string(),
-                company: "ACME".to_string(),
-            },
-        };
-
-        let mut header = Header::default();
-        header.key_id = Some("kid".to_string());
-
-        let expected_jwt = JWT::new_decoded(header.clone(), expected_claims);
-        let token = not_err!(expected_jwt.into_encoded(Secret::Bytes("secret".to_string().into_bytes())));
-        let biscuit = not_err!(token.into_decoded(Secret::Bytes("secret".to_string().into_bytes()),
-                                                  Algorithm::HS256));
-        assert_eq!(header, *not_err!(biscuit.header()));
-    }
-
-    #[test]
-    #[should_panic(expected = "InvalidToken")]
-    fn decode_token_missing_parts() {
-        let token = JWT::<PrivateClaims>::new_encoded("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9");
-        let claims = token.decode(Secret::Bytes("secret".to_string().into_bytes()),
-                                  Algorithm::HS256);
-        claims.unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "InvalidSignature")]
-    fn decode_token_invalid_signature_hs256() {
-        let token = JWT::<PrivateClaims>::new_encoded("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.\
-                                                       eyJzdWIiOiJiQGIuY29tIiwiY29tcGFueSI6IkFDTUUifQ.\
-                                                       WRONGWRONGWRONGWRONGWRONGWRONGWRONGWRONG___");
-        let claims = token.decode(Secret::Bytes("secret".to_string().into_bytes()),
-                                  Algorithm::HS256);
-        claims.unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "InvalidSignature")]
-    fn decode_token_invalid_signature_rs256() {
-        let token = JWT::<PrivateClaims>::new_encoded("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.\
-                                                       eyJzdWIiOiJiQGIuY29tIiwiY29tcGFueSI6IkFDTUUifQ.\
-                                                       WRONGWRONGWRONGWRONGWRONGWRONGWRONGWRONG___");
-        let public_key = Secret::public_key_from_file("test/fixtures/rsa_public_key.der").unwrap();
-        let claims = token.decode(public_key, Algorithm::RS256);
-        claims.unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "WrongAlgorithmHeader")]
-    fn decode_token_wrong_algorithm() {
-        let token = JWT::<PrivateClaims>::new_encoded("eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.\
-                                                       eyJzdWIiOiJiQGIuY29tIiwiY29tcGFueSI6IkFDTUUifQ.\
-                                                       pKscJVk7-aHxfmQKlaZxh5uhuKhGMAa-1F5IX5mfUwI");
-        let claims = token.decode(Secret::Bytes("secret".to_string().into_bytes()),
-                                  Algorithm::HS256);
-        claims.unwrap();
     }
 
     #[test]
